In order to make truely programmatic usage available, made cb a param of validate contracts.
