+++ lib/runner.js
@@ -1,6 +1,8 @@
+var _ = require('lodash');
 var FileHound = require('filehound');
 var path = require('path');
 var fs = require('fs');
+var formatter = require('./formatter');
 var validateContracts = require('./validator');
 
 function validateFiles(files) {
@@ -14,7 +16,23 @@ function validateFiles(files) {
     return process.exit(1);
   }
 
-  validateContracts(contracts);
+  validateContracts(contracts, function (err, results) {
+    var failures = _(results).filter('err').compact().value();
+    var totalCompleted = contracts.length;
+    var totalFailed = failures.length;
+    var totalPassed = totalCompleted - totalFailed;
+    var exitCode = (failures.length > 0) ? 1 : 0;
+
+    formatter.print({
+      results: results,
+      failures: failures,
+      totalCompleted: totalCompleted,
+      totalFailed: totalFailed,
+      totalPassed: totalPassed
+    });
+
+    process.exit(exitCode);
+  });
 }
 
 module.exports = function (files) {
